+++ protected/humhub/modules/admin/components/Controller.php
@@ -23,7 +23,7 @@ class Controller extends \humhub\components\Controller
      * @inheritdoc
      */
     public $subLayout = "@humhub/modules/admin/views/layouts/main";
-    
+
     /**
      * @var boolean if true only allows access for system admins else the access is restricted by getAccessRules()
      */
@@ -35,7 +35,7 @@ class Controller extends \humhub\components\Controller
     public function init()
     {
         $this->appendPageTitle(\Yii::t('AdminModule.base', 'Administration'));
-        return parent::init();
+        parent::init();
     }
 
     /**
@@ -44,7 +44,7 @@ class Controller extends \humhub\components\Controller
     public function behaviors()
     {
         // Workaround for module configuration actions @see getAccessRules()
-        if (Yii::$app->controller->module->id != 'admin') {
+        if ($this->module->id != 'admin') {
             $this->adminOnly = false;
         }
 
@@ -52,15 +52,21 @@ class Controller extends \humhub\components\Controller
             'acl' => [
                 'class' => AccessControl::className(),
                 'adminOnly' => $this->adminOnly,
-                'rules' => static::getAccessRules()
+                'rules' => $this->getAccessRules()
             ]
         ];
     }
 
-    public static function getAccessRules()
+    /**
+     * Returns access rules for the standard access control behavior
+     * 
+     * @see AccessControl
+     * @return array the access permissions
+     */
+    public function getAccessRules()
     {
-        // Workaround for module configuration actions
-        if (Yii::$app->controller->module->id != 'admin') {
+        // Use by default ManageModule permission, if method is not overwritten by custom module
+        if ($this->module->id != 'admin') {
             return [
                 ['permissions' => \humhub\modules\admin\permissions\ManageModules::className()]
             ];