+++ protected/humhub/modules/admin/controllers/SettingController.php
@@ -55,7 +55,7 @@ class SettingController extends Controller
     /**
      * @inheritdoc
      */
-    public static function getAccessRules()
+    public function getAccessRules()
     {
         return [
             ['permissions' => \humhub\modules\admin\permissions\ManageSettings::className()]