+++ protected/humhub/modules/admin/controllers/SpaceController.php
@@ -36,7 +36,10 @@ class SpaceController extends Controller
         return parent::init();
     }
 
-    public static function getAccessRules()
+    /**
+     * @inheritdoc
+     */
+    public function getAccessRules()
     {
         return [
             ['permissions' => [