+++ protected/humhub/modules/user/controllers/AccountController.php
@@ -13,6 +13,8 @@ use yii\web\HttpException;
 use humhub\modules\user\components\BaseAccountController;
 use humhub\modules\user\models\User;
 use humhub\modules\notification\models\forms\NotificationSettings;
+use humhub\modules\user\controllers\ImageController;
+
 /**
  * AccountController provides all standard actions for the current logged in
  * user account.
@@ -23,6 +25,9 @@ use humhub\modules\notification\models\forms\NotificationSettings;
 class AccountController extends BaseAccountController
 {
 
+    /**
+     * @inheritdoc
+     */
     public function init()
     {
         $this->setActionTitles([
@@ -213,7 +218,7 @@ class AccountController extends BaseAccountController
         if (!$user->isModuleEnabled($moduleId)) {
             $user->enableModule($moduleId);
         }
-        
+
         if (!Yii::$app->request->isAjax) {
             return $this->redirect(['/user/account/edit-modules']);
         } else {
@@ -233,7 +238,7 @@ class AccountController extends BaseAccountController
             $user->disableModule($moduleId);
         }
 
-         if (!Yii::$app->request->isAjax) {
+        if (!Yii::$app->request->isAjax) {
             return $this->redirect(['/user/account/edit-modules']);
         } else {
             Yii::$app->response->format = 'json';
@@ -366,129 +371,57 @@ class AccountController extends BaseAccountController
 
     /**
      * Crops the banner image of the user
+     * @deprecated since version 1.2
      */
     public function actionCropBannerImage()
     {
-        $model = new \humhub\models\forms\CropProfileImage();
-        $profileImage = new \humhub\libs\ProfileBannerImage($this->getUser()->guid);
-
-        if ($model->load(Yii::$app->request->post()) && $model->validate()) {
-            $profileImage->cropOriginal($model->cropX, $model->cropY, $model->cropH, $model->cropW);
-            return $this->htmlRedirect($this->getUser()->getUrl());
-        }
-
-        return $this->renderAjax('cropBannerImage', ['model' => $model, 'profileImage' => $profileImage, 'user' => $this->getUser()]);
+        return Yii::$app->runAction('/user/image/crop', ['type' => ImageController::TYPE_PROFILE_BANNER_IMAGE]);
     }
 
     /**
      * Handle the banner image upload
+     * 
+     * @deprecated since version 1.2
      */
     public function actionBannerImageUpload()
     {
-        \Yii::$app->response->format = 'json';
-
-        $model = new \humhub\models\forms\UploadProfileImage();
-        $json = array();
-
-        $files = \yii\web\UploadedFile::getInstancesByName('bannerfiles');
-        $file = $files[0];
-        $model->image = $file;
-
-        if ($model->validate()) {
-            $profileImage = new \humhub\libs\ProfileBannerImage($this->getUser()->guid);
-            $profileImage->setNew($model->image);
-
-            $json['error'] = false;
-            $json['name'] = "";
-            $json['url'] = $profileImage->getUrl();
-            $json['size'] = $model->image->size;
-            $json['deleteUrl'] = "";
-            $json['deleteType'] = "";
-        } else {
-            $json['error'] = true;
-            $json['errors'] = $model->getErrors();
+        // Ensure view file backward compatibility prior 1.2
+        if (isset($_FILES['bannerfiles'])) {
+            $_FILES['images'] = $_FILES['bannerfiles'];
         }
-
-        return ['files' => $json];
+        return Yii::$app->runAction('/user/image/upload', ['type' => ImageController::TYPE_PROFILE_BANNER_IMAGE]);
     }
 
     /**
      * Handle the profile image upload
+     * 
+     * @deprecated since version 1.2
      */
     public function actionProfileImageUpload()
     {
-        \Yii::$app->response->format = 'json';
-
-        $model = new \humhub\models\forms\UploadProfileImage();
-
-        $json = array();
-
-        $files = \yii\web\UploadedFile::getInstancesByName('profilefiles');
-        $file = $files[0];
-        $model->image = $file;
-
-        if ($model->validate()) {
-
-            $json['error'] = false;
-
-            $profileImage = new \humhub\libs\ProfileImage($this->getUser()->guid);
-            $profileImage->setNew($model->image);
-
-            $json['name'] = "";
-            $json['url'] = $profileImage->getUrl();
-            $json['size'] = $model->image->size;
-            $json['deleteUrl'] = "";
-            $json['deleteType'] = "";
-        } else {
-            $json['error'] = true;
-            $json['errors'] = $model->getErrors();
+        // Ensure view file backward compatibility prior 1.2
+        if (isset($_FILES['profilefiles'])) {
+            $_FILES['images'] = $_FILES['profilefiles'];
         }
-
-        return array('files' => $json);
+        return Yii::$app->runAction('/user/image/upload', ['type' => ImageController::TYPE_PROFILE_IMAGE]);
     }
 
     /**
      * Crops the profile image of the user
+     * @deprecated since version 1.2
      */
     public function actionCropProfileImage()
     {
-        $model = new \humhub\models\forms\CropProfileImage();
-        $profileImage = new \humhub\libs\ProfileImage($this->getUser()->guid);
-
-        if ($model->load(Yii::$app->request->post()) && $model->validate()) {
-            $profileImage->cropOriginal($model->cropX, $model->cropY, $model->cropH, $model->cropW);
-            return $this->htmlRedirect($this->getUser()->getUrl());
-        }
-
-        return $this->renderAjax('cropProfileImage', array('model' => $model, 'profileImage' => $profileImage, 'user' => $this->getUser()));
+        return Yii::$app->runAction('/user/image/crop', ['type' => ImageController::TYPE_PROFILE_IMAGE]);
     }
 
     /**
      * Deletes the profile image or profile banner
+     * @deprecated since version 1.2
      */
     public function actionDeleteProfileImage()
     {
-        \Yii::$app->response->format = 'json';
-
-        $this->forcePostRequest();
-
-        $type = Yii::$app->request->get('type', 'profile');
-
-        $json = array('type' => $type);
-
-        $image = null;
-        if ($type == 'profile') {
-            $image = new \humhub\libs\ProfileImage($this->getUser()->guid);
-        } elseif ($type == 'banner') {
-            $image = new \humhub\libs\ProfileBannerImage($this->getUser()->guid);
-        }
-
-        if ($image) {
-            $image->delete();
-            $json['defaultUrl'] = $image->getUrl();
-        }
-
-        return $json;
+        return Yii::$app->runAction('/user/image/delete', ['type' => (Yii::$app->request->get('type', 'profile') == 'profile') ? ImageController::TYPE_PROFILE_IMAGE : ImageController::TYPE_PROFILE_BANNER_IMAGE]);
     }
 
     /**