+++ protected/humhub/modules/user/widgets/ProfileHeader.php
@@ -2,7 +2,7 @@
 
 /**
  * @link https://www.humhub.org/
- * @copyright Copyright (c) 2016 HumHub GmbH & Co. KG
+ * @copyright Copyright (c) 2017 HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
  */
 
@@ -13,6 +13,7 @@ use humhub\modules\space\models\Space;
 use humhub\modules\user\models\User;
 use humhub\modules\space\models\Membership;
 use humhub\modules\friendship\models\Friendship;
+use humhub\modules\user\controllers\ImageController;
 
 /**
  * Displays the profile header of a user
@@ -45,13 +46,8 @@ class ProfileHeader extends \yii\base\Widget
             $this->user = $this->getController()->getUser();
         }
 
-        // Check if profile header can be edited
-        if (!Yii::$app->user->isGuest) {
-            if (Yii::$app->user->getIdentity()->isSystemAdmin() && Yii::$app->getModule('user')->adminCanChangeUserProfileImages) {
-                $this->isProfileOwner = true;
-            } elseif (Yii::$app->user->id == $this->user->id) {
-                $this->isProfileOwner = true;
-            }
+        if (!Yii::$app->user->isGuest && Yii::$app->user->id == $this->user->id) {
+            $this->isProfileOwner = true;
         }
 
         parent::init();
@@ -71,10 +67,8 @@ class ProfileHeader extends \yii\base\Widget
 
         $countFollowing = $this->user->getFollowingCount(User::className());
 
-        $countUserSpaces = Membership::getUserSpaceQuery($this->user)
-                ->andWhere(['!=', 'space.visibility', Space::VISIBILITY_NONE])
-                ->andWhere(['space.status' => Space::STATUS_ENABLED])
-                ->count();
+        /* @var $imageController ImageController  */
+        $imageController = new ImageController('image-controller', null, ['user' => $this->user]);
 
         return $this->render('profileHeader', array(
                     'user' => $this->user,
@@ -83,8 +77,23 @@ class ProfileHeader extends \yii\base\Widget
                     'countFriends' => $countFriends,
                     'countFollowers' => $this->user->getFollowerCount(),
                     'countFollowing' => $countFollowing,
-                    'countSpaces' => $countUserSpaces,
+                    'countSpaces' => $this->getFollowingSpaceCount(),
+                    'allowModifyProfileImage' => $imageController->allowModifyProfileImage,
+                    'allowModifyProfileBanner' => $imageController->allowModifyProfileBanner,
         ));
+    }
+
+    /**
+     * Returns the number of followed public space
+     * 
+     * @return int the follow count
+     */
+    protected function getFollowingSpaceCount()
+    {
+        return Membership::getUserSpaceQuery($this->user)
+                        ->andWhere(['!=', 'space.visibility', Space::VISIBILITY_NONE])
+                        ->andWhere(['space.status' => Space::STATUS_ENABLED])
+                        ->count();
     }
 
 }